backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
langgraph==0.2.67
sqlalchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
bcrypt==4.0.1
